Move package functions to scope
